+++ CHANGELOG.rst
@@ -4,6 +4,13 @@ Changelog
 =========
 
 
+- Add ``lower_case``. Thanks bharadwajyarlagadda_!
+- Add ``lower_first``. Thanks bharadwajyarlagadda_!
+- Add ``upper_case``. Thanks bharadwajyarlagadda_!
+- Add ``upper_first``. Thanks bharadwajyarlagadda_!
+- Add ``subtract``. Thanks bharadwajyarlagadda_!
+
+
 v3.4.5 (2016-10-16)
 -------------------
 
@@ -567,3 +574,4 @@ v0.0.0 (2014-07-22)
 
 .. _nathancahill: https://github.com/nathancahill
 .. _k7sleeper: https://github.com/k7sleeper
+.. _bharadwajyarlagadda: https://github.com/bharadwajyarlagadda